// ForceBride-Support/src/types/AdditionSubtraction.d.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import BN from "bn.js";
import { ContractOptions } from "web3-eth-contract";
import { EventLog } from "web3-core";
import { EventEmitter } from "events";
import {
  Callback,
  PayableTransactionObject,
  NonPayableTransactionObject,
  BlockType,
  ContractEventLog,
  BaseContract,
} from "./types";

interface EventOptions {
  filter?: object;
  fromBlock?: BlockType;
  topics?: string[];
}

export interface AdditionSubtraction extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): AdditionSubtraction;
  clone(): AdditionSubtraction;
  methods: {
    add(x: number | string | BN): PayableTransactionObject<void>;

    subtract(y: number | string | BN): PayableTransactionObject<void>;

    show(): NonPayableTransactionObject<string>;

    clear(): PayableTransactionObject<void>;
  };
  events: {
    allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
  };
}